+++ src/upload/catalog/language/en-gb/extension/payment/payment_gateway_cloud.php
@@ -1,6 +1,7 @@
 <?php
 $_['order_cancelled'] = 'Order was canceled.';
 $_['order_error'] = 'An error occurred during the payment process. Please try again.';
+$_['order_on_hold'] = 'Awaiting capture/void.';
 $_['button_confirm'] = 'Confirm Order';
 
 $_['redirect_text'] = 'You are being redirected. Please wait.';
@@ -11,3 +12,4 @@ $_['card_number'] = 'Card Number';
 $_['card_ccv'] = 'CCV';
 $_['card_expiry_month'] = 'Expiry Month';
 $_['card_expiry_year'] = 'Expiry Year';
+